fix: expose proto route collector API from plugin-route-analysis module
Move ArmeriaProtoRouteCollector interface into plugin-route-analysis and add
registerArmeriaGrpcProtoRoute for the PSI collector implementation in plugin.
