+++ plugin-route-analysis/src/main/kotlin/com/linecorp/intellij/plugins/armeria/explorer/ArmeriaProtoRouteCollector.kt
@@ -1,0 +1,17 @@
+package com.linecorp.intellij.plugins.armeria.explorer
+
+import com.intellij.openapi.extensions.ExtensionPointName
+import com.intellij.psi.PsiFile
+
+interface ArmeriaProtoRouteCollector {
+    fun collectFromFile(
+        file: PsiFile,
+        routes: MutableList<ArmeriaRoute>,
+        seenProtoRoutes: MutableSet<String>,
+    ): Boolean
+
+    companion object {
+        val EP: ExtensionPointName<ArmeriaProtoRouteCollector> =
+            ExtensionPointName.create("com.linecorp.intellij.armeria.protoRouteCollector")
+    }
+}
